Add manual mode to provision cmd and add timestamp field
The host_configuration.json now has a timestamp field and
the missing functionality from the old cfgtool repo is
now added back here.
